Uploaded my code for 1d GP equation
Solved the problem related to the complex number warning. But solutions are still not right : / In particular they don't keep their initial symmetry. 
Maybe the problem comes mainly from the choice of paramaters (those of the equation as well as the integration paramaters). In particular, I will improve the code to be able to choose an arbitrary interval [a,b] so as to reproduce the exact same parameters from the paper.