Observation of vortices.
Yet, the vortices are turning in the oppsosite direction than they should. One reason might be that we are calculating the gradient in the opposite direction but I am not sure it is that, I checked and it seemed to be the right direction. 
On the other side there can be some sign problem in some place that I was not able to find, yet is something that can be done.