Vertices 1.1
This is more or less the same as before, with some minor changes. 
The gradient function was fixed cause it was determining -\grad(\psi) and it was the reason why the vortices were turning around in the opposite direction. 
There are still some minor differences with the results in the paper, yet cool phenomena of formation of pair of vortices ca be observed. 
We already have fig. (1,3,5) of the paper, just missing the 2nd which is actually easy to do, is just make slices of the 3D plot of the density, and Fig.4 which is the one of the phase slip. 
After that we can move to do as Ayoub propposed and try different velocities! It would be cool!